+++ src/SimulationRunner.hh
@@ -49,6 +49,7 @@
 
 #include "network/NetworkManager.hh"
 #include "LevelManager.hh"
+#include "Barrier.hh"
 
 using namespace std::chrono_literals;
 
@@ -60,54 +61,6 @@ namespace ignition
     inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
     // Forward declarations.
     class SimulationRunnerPrivate;
-
-    class Barrier
-    {
-      public: Barrier(unsigned int _numThreads):
-                numThreads(_numThreads),
-                count(_numThreads),
-                generation(0) {}
-
-      public: bool wait()
-      {
-        if (this->cancelled)
-        {
-          return true;
-        }
-
-        std::unique_lock<std::mutex> lock(this->mutex);
-        unsigned int gen = this->generation;
-
-        if (--this->count == 0)
-        {
-          this->generation++;
-          this->count = this->numThreads;
-          this->cv.notify_all();
-          return true;
-        }
-
-        while (gen == this->generation)
-        {
-          this->cv.wait(lock);
-        }
-        return false;
-      }
-
-      public: void cancel()
-      {
-        std::unique_lock<std::mutex> lock(this->mutex);
-        this->generation++;
-        this->cancelled = true;
-        this->cv.notify_all();
-      }
-
-      private: std::mutex mutex;
-      private: std::condition_variable cv;
-      private: unsigned int numThreads;
-      private: unsigned int count;
-      private: unsigned int generation;
-      private: std::atomic<bool> cancelled = false;
-    };
 
     /// \brief Class to hold systems internally
     class SystemInternal
@@ -349,8 +302,6 @@ namespace ignition
       /// \brief Pending systems to be added to systems.
       private: std::vector<SystemPluginPtr> pendingSystems;
 
-      private: std::mutex consoleMutex;
-
       /// \brief Mutex to protect pendingSystems
       private: mutable std::mutex pendingSystemsMutex;
 
@@ -463,9 +414,16 @@ namespace ignition
       /// \brief Copy of the server configuration.
       public: ServerConfig serverConfig;
 
+      /// \brief Collection of threads running system PostUpdates
       private: std::vector<std::thread> postUpdateThreads;
-      private: std::atomic<bool> postUpdateThreadsRunning;
+
+      /// \brief Flag to indicate running status of PostUpdate threads
+      private: std::atomic<bool> postUpdateThreadsRunning{false};
+
+      /// \brief Barrier to signal beginning of PostUpdate thread execution
       private: std::unique_ptr<Barrier> postUpdateStartBarrier;
+
+      /// \brief Barrier to signal end of PostUpdate thread execution
       private: std::unique_ptr<Barrier> postUpdateStopBarrier;
 
       friend class LevelManager;