+++ src/systems/sensors/Sensors.cc
@@ -15,6 +15,9 @@
  *
  */
 
+#include <set>
+
+#include <ignition/common/Profiler.hh>
 #include <ignition/plugin/Register.hh>
 
 #include <sdf/Sensor.hh>
@@ -30,6 +33,7 @@
 #include "ignition/gazebo/components/DepthCamera.hh"
 #include "ignition/gazebo/components/GpuLidar.hh"
 #include "ignition/gazebo/components/RgbdCamera.hh"
+#include "ignition/gazebo/Events.hh"
 #include "ignition/gazebo/EntityComponentManager.hh"
 
 #include "ignition/gazebo/rendering/RenderUtil.hh"
@@ -53,10 +57,227 @@ class ignition::gazebo::systems::SensorsPrivate
   /// \brief Main rendering interface
   public: RenderUtil renderUtil;
 
+  /// \brief Unique set of senor ids
+  public: std::set<sensors::SensorId> sensorIds;
+
   /// \brief rendering scene to be managed by the scene manager and used to
   /// generate sensor data
   public: rendering::ScenePtr scene;
+
+  /// \brief Flag to indicate if worker threads are running
+  public: std::atomic<bool> running { false };
+
+  /// \brief Flag to signal if initialization should occur
+  public: bool doInit { false };
+
+  /// \brief Flag to signal if rendering update is needed
+  public: bool updateAvailable { false };
+
+  /// \brief Thread that rendering will occur in
+  public: std::thread renderThread;
+
+  /// \brief Mutex to protect rendering data
+  public: std::mutex renderMutex;
+
+  /// \brief Condition variable to signal rendering thread
+  ///
+  /// This variable is used to block/unblock operations in the rendering
+  /// thread.  For a more detailed explanation on the flow refer to the
+  /// documentation on RenderThread.
+  public: std::condition_variable renderCv;
+
+  /// \brief Connection to events::Stop event, used to stop thread
+  public: ignition::common::ConnectionPtr stopConn;
+
+  /// \brief Update time for the next rendering iteration
+  public: ignition::common::Time updateTime;
+
+  /// \brief Sensors to include in the next rendering iteration
+  public: std::vector<sensors::RenderingSensor *> activeSensors;
+
+  /// \brief Mutex to protect sensorMask
+  public: std::mutex sensorMaskMutex;
+
+  /// \brief Mask sensor updates for sensors currently being rendered
+  public: std::map<sensors::SensorId, ignition::common::Time> sensorMask;
+
+  /// \brief Wait for initialization to happen
+  private: void WaitForInit();
+
+  /// \brief Run one rendering iteration
+  private: void RunOnce();
+
+  /// \brief Top level function for the rendering thread
+  ///
+  /// This function captures all of the behavior of the rendering thread.
+  /// The behavior is captured in two phases: initialization and steady state.
+  ///
+  /// When the thread is first started, it waits on renderCv until the
+  /// prerequisites for initialization are met, and the `doInit` flag is set.
+  /// In order for initialization to proceed, rendering sensors must be
+  /// available in the EntityComponentManager.
+  ///
+  /// When doInit is set, and renderCv is notified, initialization
+  /// is performed (creating the render context and scene). During
+  /// initialization, execution is blocked for the caller of PostUpdate.
+  /// When initialization is complete, PostUpdate will be notified via
+  /// renderCv and execution will continue.
+  ///
+  /// Once in steady state, a rendering operation is triggered by setting
+  /// updateAvailable to true, and notifying via the renderCv.
+  /// The rendering operation is done in `RunOnce`.
+  ///
+  /// The caller of PostUpdate will not be blocked if there is no
+  /// rendering operation currently ongoing. Rendering will occur
+  /// asyncronously.
+  //
+  /// The caller of PostUpdate will be blocked if there is a rendering
+  /// operation currently ongoing, until that completes.
+  private: void RenderThread();
+
+  /// \brief Launch the rendering thread
+  public: void Run();
+
+  /// \brief Stop the rendering thread
+  public: void Stop();
 };
+
+//////////////////////////////////////////////////
+void SensorsPrivate::WaitForInit()
+{
+  while (!this->initialized && this->running)
+  {
+    igndbg << "Waiting for init" << std::endl;
+    std::unique_lock<std::mutex> lock(this->renderMutex);
+    // Wait to be ready for initialization or stopped running.
+    // We need rendering sensors to be available to initialize.
+    this->renderCv.wait(lock, [this]()
+    {
+      return this->doInit || !this->running;
+    });
+
+    if (this->doInit)
+    {
+      // Only initialize if there are rendering sensors
+      igndbg << "Initializing render context" << std::endl;
+      this->renderUtil.Init();
+      this->scene = this->renderUtil.Scene();
+      this->initialized = true;
+    }
+
+    this->updateAvailable = false;
+    this->renderCv.notify_one();
+  }
+  igndbg << "Rendering Thread initialized" << std::endl;
+}
+
+//////////////////////////////////////////////////
+void SensorsPrivate::RunOnce()
+{
+  std::unique_lock<std::mutex> lock(this->renderMutex);
+  this->renderCv.wait(lock, [this]()
+  {
+    return !this->running || this->updateAvailable;
+  });
+
+  if (!this->running)
+    return;
+
+  IGN_PROFILE("SensorsPrivate::RunOnce");
+  {
+    IGN_PROFILE("Update");
+    this->renderUtil.Update();
+  }
+
+  if (!this->activeSensors.empty())
+  {
+    this->sensorMaskMutex.lock();
+    // Check the active sensors against masked sensors.
+    //
+    // The internal state of a rendering sensor is not updated until the
+    // rendering operation is complete, which can leave us in a position
+    // where the sensor is falsely indicating that an update is needed.
+    //
+    // To prevent this, add sensors that are currently being rendered to
+    // a mask. Sensors are removed from the mask when 90% of the update
+    // delta has passed, which will allow rendering to proceed.
+    for (const auto & sensor : this->activeSensors)
+    {
+      // 90% of update delta (1/UpdateRate());
+      ignition::common::Time delta(0.9 / sensor->UpdateRate());
+      this->sensorMask[sensor->Id()] = this->updateTime + delta;
+    }
+    this->sensorMaskMutex.unlock();
+
+    {
+      IGN_PROFILE("PreRender");
+      // Update the scene graph manually to improve performance
+      // We only need to do this once per frame It is important to call
+      // sensors::RenderingSensor::SetManualSceneUpdate and set it to true
+      // so we don't waste cycles doing one scene graph update per sensor
+      this->scene->PreRender();
+    }
+
+    {
+      // publish data
+      IGN_PROFILE("RunOnce");
+      this->sensorManager.RunOnce(this->updateTime);
+    }
+
+    this->activeSensors.clear();
+  }
+
+  this->updateAvailable = false;
+  lock.unlock();
+  this->renderCv.notify_one();
+}
+
+//////////////////////////////////////////////////
+void SensorsPrivate::RenderThread()
+{
+  IGN_PROFILE_THREAD_NAME("RenderThread");
+
+  igndbg << "SensorsPrivate::RenderThread started" << std::endl;
+
+  // We have to wait for rendering sensors to be available
+  this->WaitForInit();
+
+  while (this->running)
+  {
+    this->RunOnce();
+  }
+  igndbg << "SensorsPrivate::RenderThread stopped" << std::endl;
+}
+
+//////////////////////////////////////////////////
+void SensorsPrivate::Run()
+{
+  igndbg << "SensorsPrivate::Run" << std::endl;
+  this->running = true;
+  this->renderThread = std::thread(&SensorsPrivate::RenderThread, this);
+}
+
+//////////////////////////////////////////////////
+void SensorsPrivate::Stop()
+{
+  igndbg << "SensorsPrivate::Stop" << std::endl;
+  std::unique_lock<std::mutex> lock(this->renderMutex);
+  this->running = false;
+
+  if (this->stopConn)
+  {
+    // Clear connection to stop additional incoming events.
+    this->stopConn.reset();
+  }
+
+  lock.unlock();
+  this->renderCv.notify_all();
+
+  if (this->renderThread.joinable())
+  {
+    this->renderThread.join();
+  }
+}
 
 //////////////////////////////////////////////////
 Sensors::Sensors() : System(), dataPtr(std::make_unique<SensorsPrivate>())
@@ -64,14 +285,18 @@ Sensors::Sensors() : System(), dataPtr(std::make_unique<SensorsPrivate>())
 }
 
 //////////////////////////////////////////////////
-Sensors::~Sensors() = default;
+Sensors::~Sensors()
+{
+  this->dataPtr->Stop();
+}
 
 //////////////////////////////////////////////////
 void Sensors::Configure(const Entity &/*_id*/,
     const std::shared_ptr<const sdf::Element> &_sdf,
     EntityComponentManager &/*_ecm*/,
-    EventManager &/*_eventMgr*/)
+    EventManager &_eventMgr)
 {
+  igndbg << "Configuring Sensors system" << std::endl;
   // Setup rendering
   std::string engineName =
       _sdf->Get<std::string>("render_engine", "ogre2").first;
@@ -80,32 +305,86 @@ void Sensors::Configure(const Entity &/*_id*/,
   this->dataPtr->renderUtil.SetEnableSensors(true,
       std::bind(&Sensors::CreateSensor, this,
       std::placeholders::_1, std::placeholders::_2));
+
+  this->dataPtr->stopConn = _eventMgr.Connect<events::Stop>(
+      std::bind(&SensorsPrivate::Stop, this->dataPtr.get()));
+
+  // Kick off worker thread
+  this->dataPtr->Run();
 }
 
 //////////////////////////////////////////////////
 void Sensors::PostUpdate(const UpdateInfo &_info,
                          const EntityComponentManager &_ecm)
 {
-  // Only initialize if there are rendering sensors
+  IGN_PROFILE("Sensors::PostUpdate");
+
   if (!this->dataPtr->initialized &&
       (_ecm.HasComponentType(components::Camera::typeId) ||
        _ecm.HasComponentType(components::DepthCamera::typeId) ||
        _ecm.HasComponentType(components::GpuLidar::typeId) ||
        _ecm.HasComponentType(components::RgbdCamera::typeId)))
   {
-    this->dataPtr->renderUtil.Init();
-    this->dataPtr->scene = this->dataPtr->renderUtil.Scene();
-    this->dataPtr->initialized = true;
+    igndbg << "Initialization needed" << std::endl;
+    std::unique_lock<std::mutex> lock(this->dataPtr->renderMutex);
+    this->dataPtr->doInit = true;
+    this->dataPtr->renderCv.notify_one();
   }
 
-  if (!this->dataPtr->initialized)
-    return;
+  if (this->dataPtr->running && this->dataPtr->initialized)
+  {
+    this->dataPtr->renderUtil.UpdateFromECM(_info, _ecm);
 
-  this->dataPtr->renderUtil.UpdateFromECM(_info, _ecm);
-  this->dataPtr->renderUtil.Update();
+    auto time = math::durationToSecNsec(_info.simTime);
+    auto t = common::Time(time.first, time.second);
 
-  auto time = math::durationToSecNsec(_info.simTime);
-  this->dataPtr->sensorManager.RunOnce(common::Time(time.first, time.second));
+    std::vector<sensors::RenderingSensor *> activeSensors;
+
+    this->dataPtr->sensorMaskMutex.lock();
+    for (auto id : this->dataPtr->sensorIds)
+    {
+      sensors::Sensor *s = this->dataPtr->sensorManager.Sensor(id);
+      sensors::RenderingSensor *rs =
+        dynamic_cast<sensors::RenderingSensor *>(s);
+
+      auto it = this->dataPtr->sensorMask.find(id);
+      if (it != this->dataPtr->sensorMask.end())
+      {
+        if (it->second <= t)
+        {
+          this->dataPtr->sensorMask.erase(it);
+        }
+        else
+        {
+          continue;
+        }
+      }
+
+      if (rs && rs->NextUpdateTime() <= t)
+      {
+        activeSensors.push_back(rs);
+      }
+    }
+    this->dataPtr->sensorMaskMutex.unlock();
+
+    if (!activeSensors.empty() ||
+        this->dataPtr->renderUtil.PendingSensors() > 0)
+    {
+      std::unique_lock<std::mutex> lock(this->dataPtr->renderMutex);
+      this->dataPtr->renderCv.wait(lock, [this] {
+        return !this->dataPtr->running || !this->dataPtr->updateAvailable; });
+
+      if (!this->dataPtr->running)
+      {
+        return;
+      }
+
+      this->dataPtr->activeSensors = std::move(activeSensors);
+      this->dataPtr->updateTime = t;
+      this->dataPtr->updateAvailable = true;
+      this->dataPtr->renderCv.notify_one();
+    }
+  }
 }
 
 //////////////////////////////////////////////////
@@ -121,6 +400,7 @@ std::string Sensors::CreateSensor(const sdf::Sensor &_sdf,
   // Create within ign-sensors
   auto sensorId = this->dataPtr->sensorManager.CreateSensor(_sdf);
   auto sensor = this->dataPtr->sensorManager.Sensor(sensorId);
+  this->dataPtr->sensorIds.insert(sensorId);
 
   if (nullptr == sensor || sensors::NO_SENSOR == sensor->Id())
   {
@@ -133,6 +413,7 @@ std::string Sensors::CreateSensor(const sdf::Sensor &_sdf,
       dynamic_cast<sensors::RenderingSensor *>(sensor);
   renderingSensor->SetScene(this->dataPtr->scene);
   renderingSensor->SetParent(_parentName);
+  renderingSensor->SetManualSceneUpdate(true);
 
   return sensor->Name();
 }